# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The business partner api dock",
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User registration",
                "operationId": "58fe5a217ef88ae75e9b530a5e5817d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "ea134134deae11eb974c719450ef8a02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "operationId": "5ebea01ab5f21df71d38f3505718a1d7",
                "responses": {
                    "204": {
                        "description": "User logged out successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "The user cart products",
                "operationId": "4a13b65ffb6dd8dc8ee2466230f5e8c2",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add product to  cart",
                "operationId": "df04e044dd4effb84d2ca0c3efe0a583",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cart/{product}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "delete product from cart",
                "operationId": "72750b1239563bcbf360b204431f5b34",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "The categories list",
                "operationId": "be248908ad1a09b8b04c532bb3d9c964",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Add category",
                "operationId": "c36ff3a00ff7abdb4abfab92ee3c0c21",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/{category}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get category item",
                "operationId": "8a64fe55c5b98d357d8d6c02d8a1d00f",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete category item",
                "operationId": "b280bce76450e7fcd7d1e3c5cc377ad4",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/upload-image": {
            "post": {
                "tags": [
                    "Image"
                ],
                "summary": "Upload image",
                "operationId": "f398511429e0cf8208c6a1cbe12b3c2d",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get product list",
                "operationId": "181f0372a31d592d30add7f1c090caec",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create or update product",
                "operationId": "a279a0126f17acbac35e69648a6b314c",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products/{product}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get product item",
                "operationId": "743474d970236485666b6c70110ef893",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete product item",
                "operationId": "d6e5ebe777077ea7d87e2a511a4b48f8",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Operations related to user authentication"
        },
        {
            "name": "Cart",
            "description": "Operations related to adding product to cart"
        },
        {
            "name": "Category",
            "description": "Operations related with categories"
        },
        {
            "name": "Image",
            "description": "Operations related with uploading images"
        },
        {
            "name": "Product",
            "description": "Operations related with products"
        }
    ]
}